+++ lib/typescript/index.d.ts
@@ -1,4 +1,4 @@
 export { Postback } from "./Postback";
 export { NativePostback } from "./NativePostback";
-export type { PostbackConfig, PostbackOptions, EventType, EventParams, AttributionResult, InstallResponse, DeviceInfo, LogLevel, TestEventResult, NativePostbackModule, } from "./types";
+export type { PostbackConfig, PostbackOptions, EventType, EventParams, AttributionResult, InstallResponse, InstallType, DeviceInfo, LogLevel, TestEventResult, NativePostbackModule, } from "./types";
 //# sourceMappingURL=index.d.ts.map